Move progress-bar-specific utilities from copy.go to new progress_bars.go
Let's try to make copy.go shorter, and consolidate progress bar code to
new copy/progress_bars.go.

First, move the code from copy.go.

Only moves unchanged code, should not change behavior.
